refactor: seller order system
