+++ src/main/java/yep/greenFire/greenfirebackend/delivery/dto/request/DeliveryAddressRequest.java
@@ -2,6 +2,7 @@ package yep.greenFire.greenfirebackend.delivery.dto.request;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
+import yep.greenFire.greenfirebackend.delivery.domain.type.AddressType;
 
 @Getter
 @RequiredArgsConstructor
@@ -16,7 +17,7 @@ public class DeliveryAddressRequest {
     private String contactNumber;
 
     private String addressZonecode;
-    private String addressType;
+    private AddressType addressType;
     private String address;
     private String addressDetail;
     private String deliveryRequest;
