+++ src/main/java/yep/greenFire/greenfirebackend/order/domain/entity/StoreOrder.java
@@ -71,6 +71,9 @@ public class StoreOrder {
             this.rejectionDate = rejectionDate;
             this.rejectionReason = rejectionReason;
         }
+    }
 
+    public void modifyStoreOrderStatus(OrderStatus orderStatus) {
+        this.orderStatus = orderStatus;
     }
 }
