+++ src/main/java/yep/greenFire/greenfirebackend/order/domain/repository/OrderRepositoryCustomImpl.java
@@ -60,6 +60,7 @@ public class OrderRepositoryCustomImpl implements OrderRepositoryCustom {
                                         GroupBy.list(Projections.constructor(StoreOrderDTO.class,
                                                 storeOrder.storeOrderCode,
                                                 storeOrder.storeCode,
+                                                store.sellerCode,
                                                 store.storeName,
                                                 storeOrder.orderStatus,
                                                 delivery.deliveryCompany,
@@ -116,6 +117,7 @@ public class OrderRepositoryCustomImpl implements OrderRepositoryCustom {
                                         GroupBy.list(Projections.constructor(StoreOrderDTO.class,
                                                 storeOrder.storeOrderCode,
                                                 storeOrder.storeCode,
+                                                store.sellerCode,
                                                 store.storeName,
                                                 storeOrder.orderStatus,
                                                 delivery.deliveryCompany,
