+++ src/main/java/yep/greenFire/greenfirebackend/order/domain/repository/StoreOrderRepository.java
@@ -2,7 +2,11 @@ package yep.greenFire.greenfirebackend.order.domain.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import yep.greenFire.greenfirebackend.order.domain.entity.StoreOrder;
+import yep.greenFire.greenfirebackend.order.domain.type.OrderStatus;
+
+import java.util.Optional;
 
 public interface StoreOrderRepository extends JpaRepository <StoreOrder, Long>, StoreOrderRepositoryCustom {
 
+    Optional<StoreOrder> findByStoreOrderCodeAndOrderStatus(Long storeOrderCode, OrderStatus orderStatus);
 }
