+++ src/main/java/yep/greenFire/greenfirebackend/order/domain/repository/StoreOrderRepositoryCustomImpl.java
@@ -58,6 +58,7 @@ public class StoreOrderRepositoryCustomImpl implements StoreOrderRepositoryCusto
                                         GroupBy.list(Projections.constructor(StoreOrderDTO.class,
                                                 storeOrder.storeOrderCode,
                                                 storeOrder.storeCode,
+                                                store.sellerCode,
                                                 store.storeName,
                                                 storeOrder.orderStatus,
                                                 delivery.deliveryCompany,
@@ -118,6 +119,7 @@ public class StoreOrderRepositoryCustomImpl implements StoreOrderRepositoryCusto
                                         GroupBy.list(Projections.constructor(StoreOrderDTO.class,
                                                 storeOrder.storeOrderCode,
                                                 storeOrder.storeCode,
+                                                store.sellerCode,
                                                 store.storeName,
                                                 storeOrder.orderStatus,
                                                 delivery.deliveryCompany,
