+++ src/main/java/yep/greenFire/greenfirebackend/order/dto/request/OrderApprovalRequest.java
@@ -2,6 +2,7 @@ package yep.greenFire.greenfirebackend.order.dto.request;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
+import yep.greenFire.greenfirebackend.order.domain.type.OrderStatus;
 
 @RequiredArgsConstructor
 @Getter
@@ -14,16 +15,6 @@ public class OrderApprovalRequest {
 
     // -- 판매자 주문 승인, 거절
 
-    private String orderStatus;
+    private OrderStatus orderStatus;
     private String rejectionReason;
-
-    // -- 판매자 배송 처리
-
-    // 배송사, 운송장 번호
-    private String deliveryCompany;
-    private String transportNumber;
-
-    // 일반 배송, 수거, 재배송
-    private String deliveryType;
-
 }
